- Backend update from Raze.
This is mainly code cleanup from setting the compiler to a stricter warning level.

+++ src/common/utility/cmdlib.cpp
@@ -248,7 +248,7 @@ bool GetFileInfo(const char* pathname, size_t *size, time_t *time)
 	bool res = _wstat64(wstr.c_str(), &info) == 0;
 #endif
 	if (!res || (info.st_mode & S_IFDIR)) return false;
-	if (size) *size = info.st_size;
+	if (size) *size = (size_t)info.st_size;
 	if (time) *time = info.st_mtime;
 	return res;
 }
@@ -359,7 +359,7 @@ FString StripExtension(const char* path)
 	src = path + strlen(path) - 1;
 
 	//
-	// back up until a . and abort on a \
+	// back up until a . and abort on a '/'
 	//
 	while (src != path && !IsSeperator(*(src - 1)))
 	{
